Work around differences in Windows and Linux keygrabbing: Allow windows to host properly
Windows' KeyGrab (keywatch) class does not work with the 'call_after_release' flag right now.
Thus, to prevent calling cycle_receiver() twice in a row, once for the global keypress down, and
once for the hooked keyboard keyup, we make sure to only call the cycle_receiver() function when
the hooked keyboard has pressed the key both down and up.